patch-set: allow user-supplied args
Allow the user to supply the list of arguments to the "git format-patch"
sub-command instead of hard-coding it to "-n". With this change it is assumed
that the last argument on the cmdline will be the rev-list, but everything
between the command and the rev-list is to be passed to "git format-patch".

Signed-off-by: Trevor Woerner <trevor.woerner@linaro.org>